through #4 EFCore Exercise

+++ Courses/Course.cs
@@ -10,7 +10,7 @@ namespace EntityFramework
     {
         public int CourseID { get; set; } 
         public string CourseName { get; set; }
-        public int CreditHours { get; set; }
+        public decimal CreditHours { get; set; }
 
         public int SubjectID { get; set; }
         public Subject Subject { get; set; }
